training/frameworks/pytorch/lightning/distributed_training/data_parallel/ddp/plugin/MNIST/training_script/requirements.txt
pytorch-lightning == 1.5.10
lightning-bolts == 0.5.0